# docs/developers/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/templates/{id}/bulk": {
      "post": {
        "description": "Create one envelope per row from a template. The call answers at once with a batch id; envelopes are created, and sent if you ask, in the background, one row at a time. A row that fails never stops the others.\n\n- rows (required): 1 to 500 entries, each `{ recipients, externalId?, metadata? }`. `recipients` has one `{ role, name, email }` for every role of the template, as when creating one envelope.\n\n- send (optional): true sends each envelope as it is created, false (the default) leaves them drafts.\n\n- message (optional): the note in every invitation; defaults to the template’s.\n\n- A row whose people do not match the template’s roles is accepted and then reported as failed with TEMPLATE_ROLE_MISMATCH; check a batch with the list and read operations.\n\n- Idempotency-Key (optional): repeating the same key and body within 24 hours returns the same batch id (Idempotency-Replayed: true) instead of a second batch.",
        "operationId": "bulkCreate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Optional. A value unique to this attempt (a UUID is ideal). Repeating the request with the same key and the same body within 24 hours returns the resource the first request created (with Idempotency-Replayed: true) instead of creating another. Without a key, every request creates a new one.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rows": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "recipients": {
                          "minItems": 1,
                          "maxItems": 50,
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "role": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 200
                              },
                              "name": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 200
                              },
                              "email": {
                                "type": "string",
                                "maxLength": 254
                              }
                            },
                            "required": [
                              "role",
                              "name",
                              "email"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "externalId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200,
                          "pattern": "^[a-zA-Z0-9_.:@-]+$"
                        },
                        "metadata": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string",
                            "maxLength": 500
                          }
                        }
                      },
                      "required": [
                        "recipients"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "message": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "send": {
                    "default": false,
                    "type": "boolean"
                  }
                },
                "required": [
                  "rows"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The batch is accepted and running; poll it for progress.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkBatchAccepted"
                }
              }
            }
          },
          "401": {
            "description": "Error codes: API_KEY_INVALID.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Error codes: API_KEY_READ_ONLY.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Error codes: TEMPLATE_NOT_FOUND.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Error codes: TEMPLATE_ARCHIVED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Error codes: BULK_TOO_LARGE, IDEMPOTENCY_KEY_MISMATCH.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Error codes: RATE_LIMITED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "session": []
          }
        ],
        "summary": "Start a bulk send",
        "tags": [
          "Templates"
        ],
        "x-idempotency": "optional",
        "x-rate-limit": "10 per 3600s per workspace"
      }
    },
    "/api/v1/bulk-batches": {
      "get": {
        "operationId": "bulkList",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Up to 100 batch summaries.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkBatchList"
                }
              }
            }
          },
          "401": {
            "description": "Error codes: API_KEY_INVALID.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Error codes: RATE_LIMITED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "session": []
          }
        ],
        "summary": "List bulk batches",
        "tags": [
          "Templates"
        ],
        "description": "The most recent batches, newest first.\n\n- No parameters.",
        "x-idempotency": "none"
      }
    },
    "/api/v1/bulk-batches/{id}": {
      "get": {
        "operationId": "bulkGet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "`status` is PROCESSING until every row has a result, then COMPLETED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkBatchDetail"
                }
              }
            }
          },
          "401": {
            "description": "Error codes: API_KEY_INVALID.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Error codes: BULK_BATCH_NOT_FOUND.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Error codes: RATE_LIMITED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "session": []
          }
        ],
        "summary": "Read a bulk batch",
        "tags": [
          "Templates"
        ],
        "description": "Progress and the outcome of every row: the envelope it created, or the error code it failed with. Rows never carry an email address.\n\n- id (path): the batchId from starting the batch.",
        "x-idempotency": "none"
      }
    },
...
  },
  "info": {
    "title": "Envelope powered by HealthProHub API",
    "description": "Server-to-server API for creating, sending and tracking documents for signature.\n\n- **Base URL:** `/api/v1` on your Envelope host.\n- **Authentication:** `Authorization: Bearer <API key>` (scheme `apiKey`). A key can call only the operations listed here; every other route needs a signed-in session.\n- **Errors:** RFC 7807 `application/problem+json` with a stable `code`.\n- **Webhooks:** see the `x-webhooks` extension and the `Webhook*` schemas.\n- **Guide:** `docs/developers/` in the repository is the same contract as a readable guide.",
    "version": "0.9.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "A server API key, created in Settings → Integrations. Keep it on your server."
      },
      "session": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "A signed-in person’s session token (the web app)."
      },
      "embedSession": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "The token an embedded editor session receives after its launch handshake."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "code": {
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "PAYLOAD_TOO_LARGE",
              "UNSUPPORTED_MEDIA_TYPE",
              "RATE_LIMITED",
              "INTERNAL_ERROR",
              "SERVICE_UNAVAILABLE",
              "EMBED_ORIGIN_NOT_ALLOWED",
              "EMBED_SESSION_INVALID",
              "EMBED_SESSION_EXPIRED",
              "EMBED_LAUNCH_USED",
              "EMBED_SCOPE_DENIED",
              "EMBED_UPLOAD_BOUND",
              "INVALID_CREDENTIALS",
              "SESSION_EXPIRED",
              "EMAIL_ALREADY_REGISTERED",
              "FORBIDDEN_ROLE",
              "LAST_OWNER",
              "INVITE_TOKEN_INVALID",
              "INVITE_TOKEN_EXPIRED",
              "PASSWORD_RESET_TOKEN_INVALID",
              "PASSWORD_RESET_TOKEN_EXPIRED",
              "CURRENT_PASSWORD_INCORRECT",
              "TWO_FACTOR_CODE_INVALID",
              "TWO_FACTOR_CHALLENGE_INVALID",
              "TWO_FACTOR_REQUIRED",
              "TWO_FACTOR_ALREADY_ENABLED",
              "TWO_FACTOR_NOT_ENABLED",
              "FILE_REQUIRED",
              "FILE_TOO_LARGE",
              "UNSUPPORTED_FILE_TYPE",
              "INVALID_PDF",
              "ENCRYPTED_PDF",
              "PAGE_LIMIT_EXCEEDED",
              "MALWARE_DETECTED",
              "INVALID_COORDINATE_SPACE",
              "RATIO_OUT_OF_RANGE",
              "FIELD_EXCEEDS_PAGE",
              "PAGE_OUT_OF_RANGE",
              "ENVELOPE_NOT_DRAFT",
              "ENVELOPE_TERMINAL",
              "ENVELOPE_EXPIRED",
              "RECIPIENT_EMAIL_TAKEN",
              "DRAFT_REVISION_MISMATCH",
              "DOCUMENT_CATEGORY_BLOCKED",
              "RECIPIENT_HAS_NO_FIELDS",
              "NOT_READY_TO_SEND",
              "REQUIRED_FIELDS_INCOMPLETE",
              "TOKEN_INVALID",
              "TOKEN_EXPIRED",
              "TOKEN_ALREADY_USED",
              "CONSENT_REQUIRED",
              "CONSENT_TEXT_CHANGED",
              "INVALID_SIGNATURE_IMAGE",
              "REMINDER_TOO_SOON",
              "DOWNLOAD_LINK_EXPIRED",
              "DOWNLOAD_RENEW_TOO_SOON",
              "ENVELOPE_ON_LEGAL_HOLD",
              "ENVELOPE_PURGED",
              "TEMPLATE_NOT_FOUND",
              "TEMPLATE_ARCHIVED",
              "TEMPLATE_NAME_TAKEN",
              "TEMPLATE_ROLE_MISMATCH",
              "BULK_TOO_LARGE",
              "BULK_BATCH_NOT_FOUND",
              "API_KEY_INVALID",
              "API_KEY_NOT_ALLOWED",
              "API_KEY_READ_ONLY",
              "WEBHOOK_URL_NOT_ALLOWED",
              "WEBHOOK_ENDPOINT_LIMIT_REACHED",
              "WEBHOOK_ENDPOINT_TOTAL_LIMIT_REACHED",
              "WEBHOOK_ENDPOINT_ACTIVE",
              "WEBHOOK_DELIVERY_NOT_REDRIVABLE",
              "IDEMPOTENCY_KEY_REQUIRED",
              "IDEMPOTENCY_KEY_MISMATCH"
            ]
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "requestId": {
            "description": "Matches the X-Request-Id header.",
            "type": "string"
          },
          "errors": {
            "description": "Per-field problems, on VALIDATION_FAILED and NOT_READY_TO_SEND.",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": {}
            }
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "additionalProperties": {},
        "description": "RFC 7807 problem details (application/problem+json)."
      },
...
      "BulkBatchSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "templateId": {
            "type": "string",
            "format": "uuid"
          },
          "templateName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PROCESSING",
              "COMPLETED"
            ]
          },
          "send": {
            "type": "boolean"
          },
          "totalRows": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "succeededRows": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "failedRows": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string",
            "description": "UTC ISO 8601 timestamp",
            "example": "2026-09-27T10:00:00.000Z"
          },
          "finishedAt": {
            "nullable": true,
            "type": "string",
            "description": "UTC ISO 8601 timestamp",
            "example": "2026-09-27T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "templateId",
          "templateName",
          "status",
          "send",
          "totalRows",
          "succeededRows",
          "failedRows",
          "createdAt",
          "finishedAt"
        ],
        "additionalProperties": {}
      },
      "BulkBatchDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "templateId": {
            "type": "string",
            "format": "uuid"
          },
          "templateName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PROCESSING",
              "COMPLETED"
            ]
          },
          "send": {
            "type": "boolean"
          },
          "totalRows": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "succeededRows": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "failedRows": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string",
            "description": "UTC ISO 8601 timestamp",
            "example": "2026-09-27T10:00:00.000Z"
          },
          "finishedAt": {
            "nullable": true,
            "type": "string",
            "description": "UTC ISO 8601 timestamp",
            "example": "2026-09-27T10:00:00.000Z"
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "rowIndex": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "PENDING",
                    "SUCCEEDED",
                    "FAILED"
                  ]
                },
                "envelopeId": {
                  "nullable": true,
                  "type": "string",
                  "format": "uuid"
                },
                "errorCode": {
                  "nullable": true,
                  "type": "string"
                }
              },
              "required": [
                "rowIndex",
                "status",
                "envelopeId",
                "errorCode"
              ],
              "additionalProperties": {}
            }
          }
        },
        "required": [
          "id",
          "templateId",
          "templateName",
          "status",
          "send",
          "totalRows",
          "succeededRows",
          "failedRows",
          "createdAt",
          "finishedAt",
          "rows"
        ],
        "additionalProperties": {}
      },
      "BulkBatchList": {
        "type": "object",
        "properties": {
          "batches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BulkBatchSummary"
            }
          }
        },
        "required": [
          "batches"
        ],
        "additionalProperties": {}
      },
      "BulkBatchAccepted": {
        "type": "object",
        "properties": {
          "batchId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "batchId"
        ],
        "additionalProperties": {}
      },
...
    }
  },
...
}